+++ src/main/java/model/User.java
@@ -2,7 +2,6 @@ package model;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 import static controller.LottoGameController.getPurchaseAmount;
 import static controller.LottoGameController.printUserLottoNumber;
@@ -30,34 +29,11 @@ public class User {
         return userLotto;
     }
 
-
     // 구매 갯수 계산 메소드
     public static int calculateCount(int cost) {
         int count = cost / LOTTO_TICKET_PRICE;
 
         return count;
-    }
-
-    // 총 당첨금액
-    public static int getTotalWinningMoney(Map<Rank, Integer> statistics) {
-        int totalWinningMoney = 0;
-
-        for (Rank rank : Rank.values()) {
-            int winningMoney = rank.getWinningMoney();
-            int amount = statistics.get(rank);
-
-            totalWinningMoney += winningMoney * amount;
-        }
-
-        return totalWinningMoney;
-    }
-
-    // 총 수익률
-    public static double getTotalYield(double winningMoney, int amount) {
-        double amountPaid = (amount * LOTTO_TICKET_PRICE);
-        double yield = winningMoney / amountPaid;
-
-        return yield;
     }
 
 }
